operation execution modified with adapters
