+++ domain/connection/services/ConnectorTypeService.py
@@ -3,7 +3,7 @@ from injector import inject
 from infrastructor.data.DatabaseSessionManager import DatabaseSessionManager
 from infrastructor.data.Repository import Repository
 from infrastructor.dependency.scopes import IScoped
-from infrastructor.exception.OperationalException import OperationalException
+from infrastructor.exceptions.OperationalException import OperationalException
 from models.dao.connection.ConnectorType import ConnectorType
