+++ infrastructor/connection/database/connectors/MssqlDbConnector.py
@@ -1,9 +1,9 @@
 import pyodbc
-from infrastructor.data.connectors.ConnectorStrategy import ConnectorStrategy
+from infrastructor.connection.database.connectors.DatabaseConnector import DatabaseConnector
 from models.configs.DatabaseConfig import DatabaseConfig
 
 
-class MssqlDbConnector(ConnectorStrategy):
+class MssqlDbConnector(DatabaseConnector):
     def __init__(self, database_config: DatabaseConfig):
         self.database_config: DatabaseConfig = database_config
         self.database_config.driver = 'ODBC Driver 17 for SQL Server'
