+++ infrastructor/connection/database/connectors/OracleDbConnector.py
@@ -1,10 +1,10 @@
 import cx_Oracle
 from injector import inject
-from infrastructor.data.connectors.ConnectorStrategy import ConnectorStrategy
+from infrastructor.connection.database.connectors.DatabaseConnector import DatabaseConnector
 from models.configs.DatabaseConfig import DatabaseConfig
 
 
-class OracleDbConnector(ConnectorStrategy):
+class OracleDbConnector(DatabaseConnector):
     @inject
     def __init__(self, database_config: DatabaseConfig):
         self.database_config = database_config
