+++ infrastructor/exceptions/OperationalException.py
